pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "pymongo>=4.12.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "requests>=2.32.3",
    "twilio>=9.5.2",
    "sendgrid>=6.11.0",
    "trafilatura>=2.0.0",
    "flask-mail>=0.10.0",
    "reportlab>=4.4.0",
    "matplotlib>=3.10.1",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
]
